requirements.txt
async-generator==1.10
attrs==22.1.0
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
gnureadline==6.3.3
h11==0.14.0
idna==3.4
iniconfig==1.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
pycodestyle==2.9.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
